refactor: 쿼리폴더명을 server=> queries로 변경
